backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.3.1
fastapi==0.124.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
motor==3.7.1
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.5
python-dotenv==1.2.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
